backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.1
dj-database-url==2.3.0
whitenoise==6.7.0
gunicorn==21.2.0
reportlab==4.1.0